requirements.txt
flask
flask-cors
torch
diffusers
pillow
accelerate